reorganized slates a little bit
