--- pype/scripts/slate/default_style.json
@@ -1,44 +1,0 @@
-{
-    "*": {
-        "font-family": "arial",
-        "font-size": 26,
-        "font-color": "#ffffff",
-        "font-bold": false,
-        "font-italic": false,
-        "bg-color": "#777777",
-        "bg-alter-color": "#666666",
-        "__alignment-vertical__values": ["left", "center", "right"],
-        "alignment-vertical": "left",
-        "__alignment-horizontal__values": ["top", "center", "bottom"],
-        "alignment-horizontal": "top",
-        "__padding__variants": [
-            "padding-top", "padding-right", "padding-bottom", "padding-left"
-        ],
-        "padding": 0,
-        "__margin__variants": [
-            "margin-top", "margin-right", "margin-bottom", "margin-left"
-        ],
-        "margin": 0,
-    },
-    "layer": {
-
-    },
-    "image": {
-
-    },
-    "text": {
-
-    },
-    "table": {
-
-    },
-    "table-item": {
-
-    },
-    "table-item-col-0": {
-
-    },
-    "#MyName": {
-
-    }
-}
