--- pype/scripts/slate/style_schema.json
@@ -1,44 +1,0 @@
-{
-    "alignment": {
-        "description": "Alignment of item",
-        "type": "string",
-        "enum": ["left", "right", "center"],
-        "example": "left"
-    },
-    "font-family": {
-        "description": "Font type",
-        "type": "string",
-        "example": "Arial"
-    },
-    "font-size": {
-        "description": "Font size",
-        "type": "integer",
-        "example": 26
-    },
-    "font-color": {
-        "description": "Font color",
-        "type": "string",
-        "regex": (
-            "^[#]{1}[a-fA-F0-9]{1}$"
-            " | ^[#]{1}[a-fA-F0-9]{3}$"
-            " | ^[#]{1}[a-fA-F0-9]{6}$"
-        ),
-        "example": "#ffffff"
-    },
-    "font-bold": {
-        "description": "Font boldness",
-        "type": "boolean",
-        "example": True
-    },
-    "bg-color": {
-        "description": "Background color, None means transparent",
-        "type": ["string", None],
-        "regex": (
-            "^[#]{1}[a-fA-F0-9]{1}$"
-            " | ^[#]{1}[a-fA-F0-9]{3}$"
-            " | ^[#]{1}[a-fA-F0-9]{6}$"
-        ),
-        "example": "#333"
-    },
-    "bg-alter-color": None,
-}
